Removing circular dependencies
