+++ external/js/cothority/src/byzcoin/index.ts
@@ -1,4 +1,4 @@
-import ByzCoinRPC, { CONFIG_INSTANCE_ID } from "./byzcoin-rpc";
+import ByzCoinRPC, { CONFIG_INSTANCE_ID, IStorage, LocalCache } from "./byzcoin-rpc";
 import ClientTransaction, { Argument, Instruction } from "./client-transaction";
 import ChainConfig from "./config";
 import Instance, { InstanceID } from "./instance";
@@ -15,4 +15,6 @@ export {
     StateChangeBody,
     Instance,
     InstanceID,
+    IStorage,
+    LocalCache,
 };
