+++ external/js/cothority/src/network/index.ts
@@ -1,12 +1,18 @@
-import * as connection from "./connection";
+import { LeaderConnection, RosterWSConnection } from "./connection";
+import { IConnection } from "./nodes";
 import { Roster, ServerIdentity, ServiceIdentity } from "./proto";
+import { setFactory, WebSocketConnection } from "./websocket";
 import { BrowserWebSocketAdapter, WebSocketAdapter } from "./websocket-adapter";
 
 export {
-    connection,
+    setFactory,
     Roster,
     ServerIdentity,
     ServiceIdentity,
     WebSocketAdapter,
     BrowserWebSocketAdapter,
+    LeaderConnection,
+    RosterWSConnection,
+    WebSocketConnection,
+    IConnection,
 };
