+++ external/js/kyber/src/pairing/gfp2.ts
@@ -194,3 +194,17 @@ export default class GfP2 {
         return `(${this.x.toHex()},${this.y.toHex()})`;
     }
 }
+
+export const xiToPMinus1Over6 = new GfP2("8669379979083712429711189836753509758585994370025260553045152614783263110636",
+    "19998038925833620163537568958541907098007303196759855091367510456613536016040");
+export const xiToPMinus1Over3 =
+    new GfP2("26098034838977895781559542626833399156321265654106457577426020397262786167059",
+        "15931493369629630809226283458085260090334794394361662678240713231519278691715");
+export const xiToPMinus1Over2 =
+    new GfP2("50997318142241922852281555961173165965672272825141804376761836765206060036244",
+        "38665955945962842195025998234511023902832543644254935982879660597356748036009");
+export const xiToPSquaredMinus1Over3 = new BN("65000549695646603727810655408050771481677621702948236658134783353303381437752");
+export const xiTo2PSquaredMinus2Over3 = new BN("4985783334309134261147736404674766913742361673560802634030");
+export const xiToPSquaredMinus1Over6 = new BN("65000549695646603727810655408050771481677621702948236658134783353303381437753");
+export const xiTo2PMinus2Over3 = new GfP2("19885131339612776214803633203834694332692106372356013117629940868870585019582",
+    "21645619881471562101905880913352894726728173167203616652430647841922248593627");
